backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI RAG pipeline backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "python-dotenv",
    "llama-index",
    "llama-index-vector-stores-postgres",
    "llama-index-retrievers-bm25",
    "llama-index-postprocessor-cohere-rerank",
    "psycopg2-binary",
    "pydantic-settings>=2.13.1",
    "llama-index-llms-groq>=0.5.0",
    "llama-index-embeddings-google-genai>=0.5.0",
    "ragas>=0.4.3",
    "datasets>=4.8.4",
    "requests>=2.33.1",
]
